portfolio_2025/API/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
bcrypt
python-jose[cryptography]
python-multipart
